Log when there are no files
I was *super* confused about what was happening when trying to address API quota issues earlier today, and it wasn't clear to me that meetings we logged "processing recording" for didn't mean we necessarily processed and uploaded an actual recording! If we already uploaded, then the recording was deleted, and the script effectively skips the meeting. But it doesn't tell you that's what it did. Now it does!