Capture and log stderr for subprocess errors (#54)
My previous fix for this in #52 failed to cover the fact that we were explicitly throwing away and not capturing stderr from the child process, so it probably would have failed to log anything useful :(

This does a little better by actually capturing that output, but only printing it if the subprocess fails. As a bonus, we exit with `sys.exit(1)` instead of by raising an error, since the stack trace at this point really wouldn't be useful.