Log when there are no files to upload (#53)
I was *super* confused about what was happening when trying to address API quota issues earlier today, and it wasn't clear to me that meetings where we logged "processing recording" didn't mean we necessarily processed and uploaded an actual recording! If we already uploaded, then the recording was deleted, and the script effectively skips the meeting. But it doesn't tell you that's what it did. Now it does!

As a bonus, the logic is a little less nested and [hopefully] and bit clearer and more up-front.

I also replaced most string concatenation with f-strings, which should read a little nicer.