First cut at fixed youtube uploading
This works, but requires you to log in via OAuth and put credentials in at runtime, which won't work for CircleCI. We need to rewrite the auth part of this.
We'll ALSO need to fix the bits after uploading that set the playlist, since those are based on the old package that seems broken.