Fix youtube uploading script (#55)
We removed the dependency on the `youtube-upload` library because it was causing errors, is no longer maintained and would've been a headache to debug. Still some refactoring to be done, but working well enough to put into production so automated uploading can start running again.

Used code from Youtube API docs to accomplish most of what we're doing and fixed up some of the docs to reflect that. 

Co-authored-by: Rob Brackett <rob@robbrackett.com>