Don't set position for auto-sorted playlists
The YouTube API requires you to set the `position` when adding videos to manually sorted playlists, but also requires you *not* to set the `position` when adding to non-manually sorted playlists. This has been causing us a lot of heartache recently when playlists have had their sorting changed.

Unfortunately, you can't detect a playlist's sorting from the API, so the approach here is to set `position`, then fall back to not setting it if we got a `manualSortRequired` error. This also has some funky parsing code since the Google API client doesn't give us back the actual `manualSortRequired` info in its error object.