Update zoom api to v2 (#63)
The zoom v1 API was deactivated. This migrates us to zoom v2 API. 

We use an up-to-date package of zoomus and also had to use the zoom API directly to delete recordings instead of zoomus, because zoomus only implements deleting all recorded files related to a meeting using the v2 API, while we still want to retain the audio and chat files for backup.