Update Zoom to use newer OAuth2 credentials (#70)
Zoom downloads/YouTube uploads broke a few months ago, and it appears to be because the way you set up credentials has changed — you now need a separate app with server-to-server OAuth credentials. This updates the `zoomus` library to a version that supports the new credential types, and behind the scenes I’ve updated the environment variables on Circle.

It looks like there are also YouTube credential issues, too, and those aren’t fixed here.