Implementing Google Drive upload support (#81)
The script can now upload to Google Drive instead of YouTube. You can still use YouTube by using the `--service youtube` CLI option.

This adds a separate authorization file (`.gdrive-upload-credentials.json.enc`) and a new config file that determines what Google Drive folders to upload to (`gdrive-locations.json.enc`). It's encoded because I’m concerned about sharing the actual IDs (maybe a future improvement would be to specify a folder path by name?). The format is:

```
{
    "default": {
        "folder": "<folder-id>",
        "subfolder_pattern": null
    },
    "ac": {
        "folder": "<folder-id>",
        "subfolder_pattern": "_{year}.AC_Meetings"
    },
    "all_edgi": {
        "folder": "<folder-id>",
        "subfolder_pattern": null
    },
    "eew": {
        "folder": "<folder-id>",
        "subfolder_pattern": "EEW {year}"
    }
}
```

Where the top-level keys correspond to a meeting type, and the values indicate the ID of the Google Drive folder to upload to. If set, an optional `subfolder_pattern` indicates a named-by-year subfolder of `folder` to upload to.