Properly support GDrive shared drives (#86)
It turns out we have to set a few extra options in order to work with shared drives. Test runs were working fine after the switch to a shared drive, but it turns out that's because we didn't make any calls that required special options until we actually had a video to upload (some commands silently fail and just return no results). Everything mostly worked fine until we had a meeting to upload, and then it started failing because of this.

More info: https://developers.google.com/workspace/drive/api/guides/enable-shareddrives