Extract Zoom utilities to separate module
This is a simplified and less ambitious version of #76. In #73, I added a bunch of error-handling code for Zoom, but it is both awkward and adds a lot of extra stuff to the recording upload script. This moves all the Zoom-specific utilities to their own module (`lib.zoom`) and attempts to smooth out the API a little bit. It's not ideal, but I think it makes the current code much more readable and manageable.