Remove vestigial `USER_TYPES` dict
This predates me maintaining this project, but it looks like this used to be used as part of a two-step lookup for the user that is our account owner. At some point we switched to using roles which lets us do it more easily, but this lookup table was left behind, even though we no longer need it.